lambdas/image_processor/requirements.txt
boto3>=1.34.0
pillow>=10.0.0
